pyproject.toml
[tool.poetry]
name = "book-service"
version = "0.1.0"
description = "testwork"
authors = ["Alex Naumov <algena75@yandex.ru>"]
readme = "README.md"
packages = [
    {include = "web"}
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.2"
uvicorn = {version = "0.29.0", extras = ["standard"]}
pydantic-settings = "2.2.1"


[tool.poetry.group.web.dependencies]
sqlalchemy = "2.0.29"
alembic = "1.13.1"
asyncpg = {version = "0.29.0", extras = ["postgres"]}
fastapi-users = {version = "13.0.0", extras = ["sqlalchemy"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
